Move unicodedata to minipal
