+++ src/coreclr/pal/src/safecrt/wcslwr_s.cpp
@@ -11,10 +11,10 @@
 *
 *******************************************************************************/
 
-#include <wctype.h>
 #include <string.h>
 #include <errno.h>
 #include <limits.h>
+#include <minipal/utf8.h>
 #include "internal_securecrt.h"
 
 #include "mbusafecrt_internal.h"
@@ -30,7 +30,7 @@ DLLEXPORT errno_t __cdecl _wcslwr_s(char16_t *string, size_t sz)
 
     for (int i = 0; string[i] != 0; i++)
     {
-        string[i] = (char16_t)towlower(string[i]);
+        string[i] = (char16_t)minipal_tolower_invariant(string[i]);
     }
 
     _FILL_STRING(string, sz, length + 1);
