+++ src/coreclr/pal/tests/palsuite/common/palsuite.h
@@ -26,6 +26,7 @@ typedef unsigned short char16_t;
 #include <minipal/utils.h>
 #include <minipal/types.h>
 #include <errno.h>
+#include <wctype.h>
 
 #define PALTEST(testfunc, testname) \
  int __cdecl testfunc(int argc, char* argv[]); \
